feat: MAT-269 매치 리스트에 명단 등록 링크 추가, 사이드바에서 미동작 링크 제거
