--- src/components/Button/index.tsx
@@ -1,16 +1,0 @@
-import { ReactNode } from 'react';
-
-import { button } from './button.css';
-
-interface ButtonProps {
-  children: ReactNode;
-  onClick?: () => void;
-}
-
-export const Button = ({ children, onClick }: ButtonProps) => {
-  return (
-    <button className={button} onClick={onClick}>
-      {children}
-    </button>
-  );
-};
