+++ src/components/index.ts
@@ -1,5 +1,4 @@
 export * from './CommonLoader';
-export * from './Button';
 export * from './PlayerList';
 export * from './StatCounterItem';
 export * from './TeamStatCounterGrid';
